refactor(pages): unify spawn into platform/, expose spawn_task from prelude
Stage 2 of Issue #4362. Consolidate platform-specific types (web_sys aliases
on WASM, SSR stubs on native) and task-spawning utilities (`spawn_task`,
`defer_yield`) into a single `crates/reinhardt-pages/src/platform/` directory
with target-dispatched submodules. The standalone `src/spawn.rs` becomes a
thin deprecation shim that delegates to `crate::platform`, preserving the
`reinhardt_pages::spawn::*` public path for external consumers while emitting
a deprecation notice.

New structure:
  crates/reinhardt-pages/src/platform.rs            cfg dispatch
  crates/reinhardt-pages/src/platform/wasm.rs       web_sys aliases + real
                                                    spawn_task / defer_yield
  crates/reinhardt-pages/src/platform/native.rs     SSR stubs + no-op
                                                    spawn_task / defer_yield

`reinhardt_pages::prelude` now re-exports `spawn_task` and `defer_yield` as
cross-target API. The existing `prelude::spawn_local` (a re-export of
`wasm_bindgen_futures::spawn_local`) is annotated `#[deprecated]` directing
callers to the new cross-target `spawn_task`. This is the only user-visible
surface change; it is non-breaking (the deprecated re-export still resolves
and compiles).

Internal call sites (callback.rs, form/component.rs, reactive/resource.rs,
reactive/hooks/{transition,async_action}.rs) and the `form!` macro codegen
in reinhardt-pages-macros are migrated to `crate::platform::*` so the
crate's own build is deprecation-warning-free under `-D warnings`.

Verified:
- cargo check -p reinhardt-pages              (native, 1m07s)
- cargo check -p reinhardt-pages --target wasm32-unknown-unknown   (18s)
- cargo clippy -p reinhardt-pages --lib -- -D warnings             (35s)
- cargo clippy -p reinhardt-pages --target wasm32-unknown-unknown \
    --lib -- -D warnings                                            (12s)
- cargo fmt --check -p reinhardt-pages -p reinhardt-pages-macros

Fixes #4365
Refs #4362
